API

API

Types

LegendDataTypes.RadwareSigcompress

Functions

Base.getindex

Base.getindex

LegendDataTypes.getunits

LegendDataTypes.read_from_properties

LegendDataTypes.readdata

LegendDataTypes.setunits!

LegendDataTypes.to_properties

LegendDataTypes.writedata

Documentation

LegendDataTypes.RadwareSigcompress — Type.

RadwareSigcompress <: AbstractArrayCodec

source

Base.getindex — Method.

getindex(input::AbstractLegendInput, key::AbstractString)
getindex(input::AbstractLegendInput, key::AbstractString, idxs::AbstractVector)

Get object at key from input.

source

Base.getindex — Method.

setindex!(output::AbstractLegendOutput, key::AbstractString)
getindex(output::AbstractLegendOutput, key::AbstractString, idxs::AbstractVector)

Get object at key from input.

source

LegendDataTypes.getunits — Function.

getunits(x)

Get the units of x, falls back to Unitful.unit(x) if no specialized method is defined for the type of x.

LEGEND I/O packages shoud add methods for the I/O-object types they handle.

source

LegendDataTypes.read_from_properties — Function.

read_from_properties(read_property::Function, src::Any, ::Type{T}) where {T<:AbstractArrayCodec}

Create a array codec of type T from properties of src, using the src-specific function read_property(src, name::Symbol, default_value) to read each property required.

Returns an value of type T.

source

LegendDataTypes.readdata — Function.

readdata(input, SomeDataType::Type)

Read a value of type SomeDataType from input.

source

LegendDataTypes.setunits! — Function.

setunits!(x)

Set the units of x.

LEGEND I/O packages will need to add methods for the I/O-object types they handle.

source

LegendDataTypes.to_properties — Function.

write_to_properties!(write_property!::Function, dest::Any, codec::AbstractArrayCodec)

Write all information required to re-create codec to dest using thes dest-specific function write_property!(dest, name::Symbol, x).

Typically returns nothing.

source

LegendDataTypes.writedata — Function.

readdata(input, x::SomeDataType)

Write a value x to output.

source